refactor: use Ponder codegen types instead of manual GraphQL interfaces
